[FEATURE](project): Added React Query to the project, added simple prototyping queries
Added prototyped queries for receive and add data to see if context is working and ssr is still applied, everything seems to work as intended. mutations are tested with AddNewButton, queries for getUser and getEntries working in EntryGrid & dashboard root page. Also added useEntries hook to make it easier to use the same query over many components if needed.